requirements.txt
asgiref==3.4.1
confusable-homoglyphs==3.2.0
distlib==0.3.4
Django==3.2.9
django-bootstrap-form==3.4
django-ckeditor==6.1.0
django-cors-headers==3.10.0
django-js-asset==1.2.2
django-markdown==0.8.4
django-nested-admin==3.4.0
django-python-code-field==1.0.0
django-registration==3.2
django-survey-and-report==1.4.0
django-tastypie==0.14.4
filelock==3.6.0
importlib-metadata==4.8.2
Markdown==3.3.4
ordereddict==1.1
packaging==21.3
Pillow==8.4.0
platformdirs==2.5.2
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-mimeparse==1.6.0
python-monkey-business==1.0.0
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tox==3.25.0
typing-extensions==3.10.0.2
Unidecode==1.3.2
virtualenv==20.14.1
zipp==3.6.0
